Split up file
