+++ plugins/spreadsheet-view/src/SpreadsheetView/components/RowMenu.tsx
@@ -7,12 +7,13 @@ import { Instance } from 'mobx-state-tree'
 import SpreadsheetModel from '../models/Spreadsheet'
 import ViewModel from '../models/SpreadsheetView'
 
-export interface Props {
+const RowMenu = observer(function ({
+  viewModel,
+  spreadsheetModel,
+}: {
   viewModel: Instance<typeof ViewModel>
   spreadsheetModel: Instance<typeof SpreadsheetModel>
-}
-
-const RowMenu = observer(({ viewModel, spreadsheetModel }: Props) => {
+}) {
   const currentRowMenu = spreadsheetModel.rowMenuPosition
   const { setRowMenuPosition } = spreadsheetModel
